requirements.txt
numpy
opencv-python
scipy
Pillow
pytorch
progressbar2
matplotlib